fkj_server/requirements.txt
pyserial
pywin32
pynput